# openapi/opss-api.yaml
openapi: 3.0.0
info:
  title: opss-api
  version: 1.0.0
paths:
  /cats/:
    post:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
      parameters: []
      tags:
        - 梦想家/小伙伴/猫咪管理
      summary: 添加一只猫咪
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: number
                breed:
                  type: string
    get:
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          example: "Bearer your_actual_token_here"
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cat'
      tags:
        - 梦想家/小伙伴/猫咪管理
      summary: 获取所有猫咪
  /cats/:id:
    get:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
      parameters:
        - name: id
          in: path
          schema:
            type: string
      tags:
        - 梦想家/小伙伴/猫咪管理
      summary: 获取一只猫咪
    put:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
      parameters:
        - name: id
          in: path
          schema:
            type: string
      tags:
        - 梦想家/小伙伴/猫咪管理
      summary: 更新一只猫咪
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCateDtoInterface'
    delete:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                anyOf:
                  - type: boolean
                    const: false
                  - type: boolean
                    const: true
      parameters:
        - name: id
          in: path
          schema:
            type: string
      tags:
        - 梦想家/小伙伴/猫咪管理
      summary: 领养一只猫咪
  /dog/:
    post:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  age:
                    type: number
                  breed:
                    type: string
                  id:
                    type: number
                required:
                  - name
                  - age
                  - breed
                  - id
      parameters: []
      tags:
        - 梦想家/小伙伴/旺财管理
      summary: 添加一只旺财
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: number
                breed:
                  type: string
    get:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    age:
                      type: number
                    breed:
                      type: string
                    id:
                      type: number
                  required:
                    - name
                    - age
                    - breed
                    - id
      parameters: []
      tags:
        - 梦想家/小伙伴/旺财管理
      summary: 获取所有旺财
  /dog/:id:
    get:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  age:
                    type: number
                  breed:
                    type: string
                  id:
                    type: number
                required:
                  - name
                  - age
                  - breed
                  - id
      parameters:
        - name: id
          in: path
          schema:
            type: string
      tags:
        - 梦想家/小伙伴/旺财管理
      summary: 获取一只旺财
    delete:
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
      parameters:
        - name: id
          in: path
          schema:
            type: string
      tags:
        - 梦想家/小伙伴/旺财管理
      summary: 领养一只旺财
components:
  schemas:
    Cat:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: number
      required:
        - id
        - name
        - age
        - id
        - name
        - age
        - id
        - name
        - age
        - id
        - name
        - age
    UpdateCateDtoInterface:
      type: object
      properties:
        name:
          type: string
        age:
          type: number
        breed:
          type: string
      required:
        - name
        - age
        - breed